Worked on the temperature systematics fitter.
starsdb:
Added o_abund_nt and c_abund_nt to save the non-corrected abundance
Also if mkdb detects the table already exists, kill it.

Modified postfit to work with the databases as opposed to stars structure

Created a new postfit.applytfit that updates the _abund columns to hold
the temperature-corrected abundance.