Made the number of bins a variable.
Instead of having the number of bins be constant for all trial
periods, I have a parameter `ntbin` be the expected number of bins for
a transit.

IMPORTANT!!!
Since the number of bins changes depending on the trial period. mdt in
blswrap no longer works.  This can be modified by computing the phase
of ingress inside the trial period loop.