Changes to the Harmonic identifier
Before it only compared P to 2 P.  Now it compares P to a range of
harmonic specified in the harmL array at the beginning of the module.
harm compares X^2 between the input transit parameter and a harmonic.
The photometry used to compute X^2 is the union between the
fundemental and the harmonic.